refactor: help/version commands don't return error (#21)
help/version commands used to return errors. There was an awkward
switch/case for checking for command names in ArgumentError.

This has been removed and now they implement the SingleExecCommand
interface, which executes a single command without reading from STDIN.
This interface is supposed to be used for commands that just return
something such as "version" and "help".